+++ src/share/grabbag/cuesheet.c
@@ -584,7 +584,7 @@ void grabbag__cuesheet_emit(FILE *file, const FLAC__StreamMetadata *cuesheet, co
 #ifdef _MSC_VER
 				fprintf(file, "%I64u\n", track->offset + index->offset);
 #else
-				fprintf(file, "%llu\n", track->offset + index->offset);
+				fprintf(file, "%llu\n", (unsigned long long)(track->offset + index->offset));
 #endif
 		}
 	}
@@ -593,7 +593,7 @@ void grabbag__cuesheet_emit(FILE *file, const FLAC__StreamMetadata *cuesheet, co
 	fprintf(file, "REM FLAC__lead-in %I64u\n", cs->lead_in);
 	fprintf(file, "REM FLAC__lead-out %u %I64u\n", (unsigned)cs->tracks[track_num].number, cs->tracks[track_num].offset);
 #else
-	fprintf(file, "REM FLAC__lead-in %llu\n", cs->lead_in);
-	fprintf(file, "REM FLAC__lead-out %u %llu\n", (unsigned)cs->tracks[track_num].number, cs->tracks[track_num].offset);
+	fprintf(file, "REM FLAC__lead-in %llu\n", (unsigned long long)cs->lead_in);
+	fprintf(file, "REM FLAC__lead-out %u %llu\n", (unsigned)cs->tracks[track_num].number, (unsigned long long)cs->tracks[track_num].offset);
 #endif
 }
